requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
PyYAML==6.0
langchain==0.0.197
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.33.1
torch==2.0.1
pydantic==1.10.7
pdfminer.six==20221105
